issue #8 - Remote is rolledback when datasets are sent
When sending a zfs dataset incrementally, if zfs detects that the
remote dataset has changed in any way, the incremental send will
fail.

The -F flag for 'zfs recv' will force the rollback of the dataset
when this happens, allowing the incremental send to proceed.

The problem is, the -F flag also nukes any unrelated snapshots that
were made on that dataset, causing you to lose remote versioning
of your data.

The fix requires the following:
  * Stop re-mounting the dataset after the send has completed. this
    is what caused the necessity of the -F in the first place.
  * Remove the '-F' flag from 'zfs recv'

Further, the README was updated to include info on how to create
the remote dataset such that is unmountable, serving as a further
precaution against invalidated incremental snapshots