Add mode support for tags
Tag modes allow you to change how a tag is applied to a dataset.
This commit adds a new mode, 'ignore'. When a dataset is tagged
with a tag in the ignore mode, sends of that tag will explicitly
not send that dataset, even if a parent dataset is set to be sent.
Additionally, all children of that ignored dataset will be ignored.

This means you can easily tag a root dataset to be sent, but ignore
only specific datasets during a send operation (/var and /tmp maybe?)

Future modes might be added at a later date.