Remove the '-t' argument to 'create'
The target parameter isn't necessarily a universal concept for all send
schemes. For example, in the future, sending a stream to a tar file
wouldn't really need a target, rather, a filename.

This commit moves the target parameter to 'create' into the options
list for ssh (the only supported scheme for now) and moves the concept
of the 'tag target' out of the base send code and into the helper functions.