Add 'untag' command
Use 'untag' to remove a tag set on a dataset. Note that this does NOT
make it so that dataset won't be sent via this tag. Rather, this
just removes any explicit tagging set directly on that dataset.

So, for example, if the dataset untagged has a parent dataset that
has the same tag, the untagged dataset will likely still be sent
with that tag.

To explicitly block a dataset from being sent for a tag, use the
following command:

'zshadow tag pool/dataset/name tag-name -m ignore